requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
passlib[argon2,bcrypt]==1.7.4
python-dotenv==1.2.1
python-multipart==0.0.21
pydantic==2.12.4
pydantic-settings==2.12.0
email-validator==2.3.0
